[chore/#19] 패키지 변경 및 이동
