--- app/src/main/java/sopt/org/millie/presentation/bookdetail/DummyScreen1.kt
@@ -1,1 +1,0 @@
-package sopt.org.millie.presentation.bookdetail
