--- app/src/main/java/sopt/org/millie/presentation/home/model/BookCategoryModel.kt
@@ -1,11 +1,0 @@
-package sopt.org.millie.presentation.home.model
-
-import androidx.compose.runtime.Immutable
-
-@Immutable
-data class BookCategoryModel(
-    val categoryId: Long,
-    val title: String,
-    val description: String,
-    val imageUrl: String,
-)
