--- app/src/main/java/sopt/org/millie/presentation/home/model/RankingModel.kt
@@ -1,9 +1,0 @@
-package sopt.org.millie.presentation.home.model
-
-import androidx.annotation.DrawableRes
-
-data class RankingModel(
-    val rank: Int,
-    val bookTitle: String,
-    @DrawableRes val rankingIcon: Int,
-)
