+++ app/src/main/java/sopt/org/millie/presentation/search/bookdetail/component/tag/BookDetailLikedTag.kt
@@ -1,0 +1,67 @@
+package sopt.org.millie.presentation.search.bookdetail.component.tag
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.res.vectorResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import sopt.org.millie.R
+import sopt.org.millie.core.designsystem.theme.MillieTheme
+
+@Composable
+fun BookDetailLikedTag(
+    onClick: () -> Unit,
+    likedNum: Int,
+    modifier: Modifier = Modifier,
+) {
+    Row(
+        modifier = modifier
+            .clip(shape = RoundedCornerShape(16.dp))
+            .background(color = MillieTheme.colors.lightGray1)
+            .padding(horizontal = 10.dp, vertical = 5.dp),
+        verticalAlignment = Alignment.CenterVertically,
+        horizontalArrangement = Arrangement.spacedBy(3.dp),
+    ) {
+        Icon(
+            imageVector = ImageVector.vectorResource(id = R.drawable.ic_detail_heart_small),
+            contentDescription = null,
+            tint = Color.Unspecified,
+            modifier = Modifier.clickable(onClick = onClick),
+        )
+
+        Text(
+            text = "좋아요",
+            color = MillieTheme.colors.black,
+            style = MillieTheme.typography.body.caption3,
+        )
+
+        Text(
+            text = likedNum.toString(),
+            color = MillieTheme.colors.black,
+            style = MillieTheme.typography.title.subHead1,
+        )
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun Preview() {
+    MillieTheme {
+        BookDetailLikedTag(
+            onClick = {},
+            likedNum = 724,
+        )
+    }
+}
