+++ app/src/main/java/sopt/org/millie/presentation/search/home/HomeUiState.kt
@@ -1,0 +1,17 @@
+package sopt.org.millie.presentation.search.home
+
+import androidx.compose.runtime.Immutable
+import sopt.org.millie.R
+import sopt.org.millie.core.util.UiState
+import sopt.org.millie.presentation.search.home.constants.HomeConstants
+import sopt.org.millie.presentation.search.home.model.BookCategoryModel
+import sopt.org.millie.presentation.search.home.model.RankingModel
+
+@Immutable
+data class HomeUiState(
+    val searchKeyword: String = "",
+    val tabs: List<String> = HomeConstants.HOME_TABS,
+    val selectedTab: String = HomeConstants.DEFAULT_TAB,
+    val rankingItem: RankingModel = RankingModel(rank = 2, bookTitle = "동화", rankingIcon = R.drawable.ic_home_ranking_down),
+    val categoryList: UiState<List<BookCategoryModel>> = UiState.Loading,
+)
