+++ app/src/main/java/sopt/org/millie/presentation/search/home/model/RankingModel.kt
@@ -1,0 +1,9 @@
+package sopt.org.millie.presentation.search.home.model
+
+import androidx.annotation.DrawableRes
+
+data class RankingModel(
+    val rank: Int,
+    val bookTitle: String,
+    @DrawableRes val rankingIcon: Int,
+)
