+++ app/src/main/java/sopt/org/millie/presentation/search/search/component/book/SearchWandokRate.kt
@@ -1,0 +1,67 @@
+package sopt.org.millie.presentation.search.search.component.book
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.res.vectorResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import sopt.org.millie.R
+import sopt.org.millie.core.designsystem.theme.MillieTheme
+
+@Composable
+fun SearchWandokRate(
+    rate: Int,
+    minute: Int,
+    modifier: Modifier = Modifier,
+) {
+    Row(
+        modifier = modifier,
+        horizontalArrangement = Arrangement.spacedBy(2.dp),
+        verticalAlignment = Alignment.CenterVertically,
+    ) {
+        Icon(
+            imageVector = ImageVector.vectorResource(id = R.drawable.ic_result_wandok),
+            contentDescription = null,
+            modifier = Modifier.size(15.dp),
+            tint = MillieTheme.colors.gray3,
+        )
+
+        Spacer(modifier = Modifier.width(2.dp))
+
+        Text(
+            text = "$rate%",
+            color = MillieTheme.colors.gray3,
+            style = MillieTheme.typography.body.body1,
+        )
+
+        Text(
+            text = "|",
+            color = MillieTheme.colors.gray3,
+            style = MillieTheme.typography.body.body1,
+        )
+
+        Text(
+            text = "${minute}분",
+            color = MillieTheme.colors.gray3,
+            style = MillieTheme.typography.body.body1,
+        )
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun SearchWandokRatePreview() {
+    SearchWandokRate(
+        rate = 88,
+        minute = 247,
+    )
+}
